Fixes to improve PWA object (#74,#65,#63)
