+++ lib/images.js
@@ -59,7 +59,7 @@ function fetchAndSave(url, destFile) {
 function saveImage(response, destFile) {
   return new Promise((resolve, reject) => {
     if (response.status !== 200) {
-      reject(new Error('Bad Response'));
+      reject(new Error('Bad Response loading image: ' + response.status));
     }
 
     const contentType = response.headers.get('Content-Type');
