+++ views/pwas/view.hbs
@@ -47,7 +47,7 @@
 
         <div class="pwabox-body">
           <h3>{{pwa.name}}</h4>
-          <p>{{pwa.manifestData.description}}</p>
+          <p>{{pwa.description}}</p>
           {{#if pwa.absoluteStartUrl}}
           <h4><a href="{{pwa.absoluteStartUrl}}">Homepage</a></h4>
           {{/if}}
